pyproject.toml
[project]
name = "streamo"
version = "0.0.0"
readme = "README.md"
requires-python = "==3.10.19"
dependencies = [
  "accelerate==1.12.0",
  "addict==2.4.0",
  "aiohttp==3.13.3",
  "attrdict==2.0.1",
  "binpacking==2.0.1",
  "charset_normalizer==3.4.4",
  "cpm_kernels==1.0.11",
  "dacite==1.9.2",
  "datasets==3.6.0",
  "einops==0.8.2",
  "fastapi==0.128.0",
  "gradio==6.5.1",
  "google-cloud-storage==3.4.1",
  "importlib_metadata==8.7.1",
  "jieba==0.42.1",
  "json_repair==0.55.2",
  "matplotlib==3.10.8",
  "modelscope==1.34.0",
  "nltk==3.9.2",
  "numpy==2.2.6",
  "openai==2.16.0",
  "oss2==2.19.1",
  "pandas==2.3.3",
  "peft==0.17.1",
  "pillow==12.1.0",
  "PyYAML==6.0.3",
  "requests==2.32.5",
  "rouge==1.0.1",
  "safetensors==0.7.0",
  "scipy==1.15.3",
  "sentencepiece==0.2.1",
  "simplejson==3.20.2",
  "sortedcontainers==2.4.0",
  "tensorboard==2.20.0",
  "tiktoken==0.12.0",
  "tqdm==4.67.2",
  "transformers==4.57.6",
  "transformers_stream_generator==0.0.5",
  "trl==0.23.1",
  "uvicorn==0.40.0",
  "zstandard==0.25.0",
]
